bug fixing

+++ src/gfx.cpp
@@ -95,7 +95,7 @@ void Gfx::Initialize() {
     glBufferData(GL_ARRAY_BUFFER, sizeof(GLfloat) * 4 * 8 * MAX_QUADS, emptyData, GL_DYNAMIC_DRAW);
     glBindBuffer(GL_ARRAY_BUFFER, NULL);
 
-    GLuint* indexData = new GLuint[6 * MAX_QUADS];
+    GLushort* indexData = new GLushort[6 * MAX_QUADS];
     for (int i = 0; i < MAX_QUADS; i++) {
         indexData[i * 6 + 0] = i * 4 + 0;
         indexData[i * 6 + 1] = i * 4 + 1;
@@ -106,7 +106,7 @@ void Gfx::Initialize() {
     }
 
     glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, ibo);
-    glBufferData(GL_ELEMENT_ARRAY_BUFFER, sizeof(GLuint) * 6 * MAX_QUADS, indexData, GL_STATIC_DRAW);
+    glBufferData(GL_ELEMENT_ARRAY_BUFFER, sizeof(GLushort) * 6 * MAX_QUADS, indexData, GL_STATIC_DRAW);
     glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, NULL);
 
     Gfx::SetSize(1.0f, 1.0f);
